+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redirect user to edit wellness todo item when tapping on calendar item with reminder [#1976](https://github.com/rokwire/illinois-app/issues/1976).
 - Acknowledged new parameters of 'report/abuse' API of Groups BB [#2083](https://github.com/rokwire/illinois-app/issues/2083).
 - Improve Semantics for WellnessRings widgets [#2023](https://github.com/rokwire/illinois-app/issues/2023).
+- Improve Semantics for 8 dimensions of wellness diagram widgets [#2027](https://github.com/rokwire/illinois-app/issues/2027).
 
 ## [4.1.2] - 2022-07-25
 ### Added